requirements.txt
Django==1.6.5
argparse==1.2.1
djangorestframework==2.3.13
wsgiref==0.1.2
South==0.8.4
psycopg2==2.5.3
celery==3.1.12
django-celery==3.1.10
django-extensions==1.3.8
django-crispy-forms==1.4.0
youtube-dl==2014.06.19
Werkzeug==0.9.6
